pyproject.toml
[tool.poetry]
name = "ml"
version = "0.1.0"
description = "Diffusion machine learning."
authors = ["Korl Zeleznikar <korl.zeleznikar@cern.ch>"]

[tool.poetry.dependencies]
python = "^3.13"

# ml
torch = "2.7.0"
torchmetrics = "1.4.1"
lightning = "2.4.0"
mlflow = "3.3.2"
scikit-learn = "1.5.1"
torchinfo = "1.8.0"
pyarrow = ">=18.0.0"

# numeric
numpy = "2.2"
pandas = ">=2.2.3,<2.4"
scipy = "1.14.1"
sympy = ">=1.13.3,<1.15"
tables = "3.10.1"
uncertainties = "3.2.2"
numba = "0.61.2"
llvmlite = "0.44.0"

# plotting
matplotlib = "3.9.2"
seaborn = "0.13.2"
corner = "2.2.2"
plothist = ">=1.2.6"

# utils
h5py = "3.14.0"
hydra-core = "1.3.2"
requests = ">=2.28.1"
tqdm = ">=4.65.0"
rich = ">=13.7.0"

# graphs
torch-geometric = "2.5.3"
pygraphviz = "1.13"
pydot = "3.0.1"
networkx = "3.3"
hep-ml = "^0.8.0"
mplhep = "^0.4.1"
colormaps = "^0.4.2"

[tool.poetry.group.dev.dependencies]
# jupyter
ipython = ">=8.21.0"
ipywidgets = ">=8.1.1"
jupyterlab = ">=4.0.11"
mkdocs = "^1.6.0"
mkdocs-material = "^9.5.33"
mkdocstrings = {extras = ["python"], version = "^0.25.2"}
pytkdocs = {version = "^0.16.1", extras = ["numpy-style"]}
mkdocs-autorefs = "^1.1.0"
pre-commit = "^3.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
